requirements.txt
# AWS Bedrock
boto3>=1.34.0
botocore>=1.34.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# HTTP and API
httpx>=0.26.0
aiohttp>=3.9.0

# XML/Atom parsing for arXiv
feedparser>=6.0.10
xmltodict>=0.13.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# CORS
fastapi-cors>=0.0.6

# Date handling
python-dateutil>=2.8.2

# Logging
structlog>=24.1.0
